python/liveai-service/requirements.txt
websockets==14.1
google-auth==2.37.0
certifi==2024.12.14
requests==2.32.3
google-genai==1.24.0
aiohttp==3.11.11
python-dotenv==1.0.1
google-api-python-client==2.122.0
google-auth-oauthlib==1.2.0
google-cloud-secret-manager==2.24.0
rocketmq-python-client==5.0.6
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation-logging==0.57b0